Ensure we don't post to unauthorized channels
I'd like to take this commit to say I'm not very happy with how this
code is organized. The next chunk of commits should almost exclusively
be tests so I can get to a good coverage percentage and THEN start
screwing around with a better organization. This is not ideal, but at
the end of the day this whole thing is kinda gimmicky to begin with.