Refactor some smaller functions, handle no internet connectivity
Currently travelling and have no actual internet reception, so
I kept commenting out the init function because I don't have internet.
Then I kinda started thinking I should actually handle that and express
to the user that there is a connectivity issue. Also that shouldn't
hold up testing, which it was doing before all this, so there.

I'm not writing off the likelihood that this will be removed later.